feat: Add option to manually backup and show results in snackbar

+++ app/src/main/java/com/simplyteam/simplybackup/presentation/viewmodels/main/ConnectionOverviewViewModel.kt
@@ -1,17 +1,24 @@
 package com.simplyteam.simplybackup.presentation.viewmodels.main
 
+import android.content.Context
 import android.content.Intent
+import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.compose.foundation.lazy.LazyListState
+import androidx.compose.material.SnackbarHostState
+import androidx.compose.material.SnackbarResult
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import com.simplyteam.simplybackup.R
 import com.simplyteam.simplybackup.data.models.Connection
+import com.simplyteam.simplybackup.data.receiver.BackupReceiver
 import com.simplyteam.simplybackup.data.repositories.ConnectionRepository
 import com.simplyteam.simplybackup.data.services.SchedulerService
 import com.simplyteam.simplybackup.data.services.search.ConnectionSearchService
 import com.simplyteam.simplybackup.data.utils.ActivityUtil.StartActivityWithAnimation
 import com.simplyteam.simplybackup.presentation.activities.ConnectionConfigurationActivity
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.launch
 import timber.log.Timber
 import javax.inject.Inject
 
@@ -23,6 +30,7 @@ class ConnectionOverviewViewModel @Inject constructor(
 ) : ViewModel() {
 
     val ListState = LazyListState()
+    val SnackbarHostState = SnackbarHostState()
 
     fun GetConnections() = _connectionSearchService.FilteredItems
 
@@ -36,11 +44,40 @@ class ConnectionOverviewViewModel @Inject constructor(
         _connectionSearchService.Search("")
     }
 
-    suspend fun DeleteConnection(connection: Connection) {
+    fun DeleteConnection(
+        context: Context,
+        connection: Connection
+    ) {
         try {
-            _connectionRepository.RemoveConnection(connection)
+            viewModelScope.launch {
+                _connectionRepository.UpdateConnection(
+                    connection.apply {
+                        TemporarilyDeleted = true
+                    }
+                )
 
-            _schedulerService.CancelBackup(connection)
+                val result = SnackbarHostState.showSnackbar(
+                    context.getString(
+                        R.string.ConnectionRemoved,
+                        connection.Name
+                    ),
+                    context.getString(
+                        R.string.Undo
+                    )
+                )
+
+                if (result == SnackbarResult.Dismissed) {
+                    _connectionRepository.RemoveConnection(connection)
+
+                    _schedulerService.CancelBackup(connection)
+                } else {
+                    _connectionRepository.UpdateConnection(
+                        connection.apply {
+                            TemporarilyDeleted = false
+                        }
+                    )
+                }
+            }
         } catch (ex: Exception) {
             Timber.e(ex)
         }
@@ -65,5 +102,50 @@ class ConnectionOverviewViewModel @Inject constructor(
         context.StartActivityWithAnimation(
             intent
         )
+    }
+
+    suspend fun RunBackup(
+        context: Context,
+        connection: Connection
+    ) {
+        try {
+            val intent = CreateIntent(
+                context,
+                connection
+            )
+
+            context.sendBroadcast(intent)
+
+            SnackbarHostState.showSnackbar(
+                context.getString(
+                    R.string.BackupStarted,
+                    connection.Name
+                )
+            )
+        } catch (ex: Exception) {
+            Timber.e(ex)
+        }
+    }
+
+    private fun CreateIntent(
+        context: Context,
+        connection: Connection
+    ): Intent {
+        val intent = Intent(
+            context,
+            BackupReceiver::class.java
+        )
+
+        val bundle = Bundle()
+        bundle.putSerializable(
+            "Connection",
+            connection
+        )
+        intent.putExtra(
+            "Bundle",
+            bundle
+        )
+
+        return intent
     }
 }
